requirements.txt
alembic==1.4.3
aniso8601==6.0.0
appdirs==1.4.4
asgiref==3.2.10
astroid==2.2.5
atomicwrites==1.4.0
attrs==20.2.0
Authlib==0.14.3
awscli==1.18.159
Babel==2.8.1
backcall==0.2.0
botocore==1.18.18
certifi==2020.6.20
cffi==1.14.3
chardet==3.0.4
click==7.1.2
colorama==0.4.4
cryptography==3.1.1
decorator==4.4.2
distlib==0.3.1
dj-database-url==0.5.0
Django==3.1.2
django-heroku==0.3.1
docutils==0.15.2
dominate==2.6.0
ecdsa==0.13.2
filelock==3.0.12
flake8==3.8.4
Flask==1.1.2
Flask-Cors==3.0.9
Flask-Migrate==2.5.3
Flask-Moment==0.10.0
Flask-RESTful==0.3.7
flask-restplus==0.13.0
Flask-Script==2.0.6
Flask-SQLAlchemy==2.4.4
flask-swagger-ui==3.36.0
Flask-WTF==0.14.3
future==0.17.1
gunicorn==20.0.4
idna==2.10
importlib-metadata==2.0.0
inflect==4.1.0
iniconfig==1.1.1
ipykernel==5.3.4
ipython==7.17.0
ipython-genutils==0.2.0
isort==4.3.18
itsdangerous==1.1.0
jedi==0.17.2
Jinja2==2.11.2
jmespath==0.10.0
jsonschema==3.2.0
jupyter-client==6.1.6
jupyter-core==4.6.3
lazy-object-proxy==1.4.0
Mako==1.1.3
MarkupSafe==1.1.1
mccabe==0.6.1
packaging==20.4
parso==0.7.1
pickleshare==0.7.5
pluggy==0.13.1
prompt-toolkit==3.0.5
psycopg2==2.8.6
psycopg2-binary==2.8.6
py==1.9.0
pyasn1==0.4.8
pycodestyle==2.6.0
pycparser==2.20
pycryptodome==3.3.1
pyflakes==2.2.0
Pygments==2.6.1
PyJWT==1.7.1
pylint==2.3.1
pyparsing==2.4.7
pyrsistent==0.17.3
pytest==6.1.1
python-dateutil==2.8.1
python-dotenv==0.14.0
python-editor==1.0.4
python-jose-cryptodome==1.3.2
pytz==2019.1
pywin32==228
PyYAML==5.3.1
pyzmq==19.0.2
requests==2.24.0
rsa==4.5
s3transfer==0.3.3
six==1.15.0
sqlacodegen==2.3.0
SQLAlchemy==1.3.20
sqlparse==0.4.1
toml==0.10.1
tornado==6.0.4
traitlets==4.3.3
typed-ast==1.3.5
urllib3==1.25.10
virtualenv==20.0.31
visitor==0.1.3
wcwidth==0.2.5
Werkzeug==1.0.1
whitenoise==5.2.0
wrapt==1.11.1
WTForms==2.3.3
zipp==3.3.1
